update namespacing and add xsds
