+++ lib/aca_entities/fdsh/pvc/h43/xsd/request/XMLschemas/exchange/ExchangeModel.xsd
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<xsd:schema xmlns:exch="http://pvcreq.dsh.cms.gov/exchange/1.0" xmlns:ext="http://pvcreq.dsh.cms.gov/extension/1.0" xmlns:i="http://niem.gov/niem/appinfo/2.0" xmlns:i2="http://niem.gov/niem/appinfo/2.1" xmlns:niem-xsd="http://niem.gov/niem/proxy/xsd/2.0" xmlns:s="http://niem.gov/niem/structures/2.0" xmlns:xsd="http://www.w3.org/2001/XMLSchema" attributeFormDefault="qualified" elementFormDefault="qualified" targetNamespace="http://pvcreq.dsh.cms.gov/exchange/1.0" version="1">
+  <xsd:annotation>
+    <xsd:appinfo>
+      <i:ConformantIndicator>true</i:ConformantIndicator>
+    </xsd:appinfo>
+    <xsd:documentation source="">An exchange model for submitting one transmission from a batch, with multiple attachments.</xsd:documentation>
+  </xsd:annotation>
+  <xsd:import namespace="http://niem.gov/niem/appinfo/2.0" schemaLocation="../../XMLschemas/subset/niem/appinfo/2.0/appinfo.xsd"/>
+  <xsd:import namespace="http://niem.gov/niem/appinfo/2.1" schemaLocation="../../XMLschemas/subset/niem/appinfo/2.1/appinfo.xsd"/>
+  <xsd:import namespace="http://niem.gov/niem/proxy/xsd/2.0" schemaLocation="../../XMLschemas/subset/niem/proxy/xsd/2.0/xsd.xsd"/>
+  <xsd:import namespace="http://niem.gov/niem/structures/2.0" schemaLocation="../../XMLschemas/subset/niem/structures/2.0/structures.xsd"/>
+  <xsd:import namespace="http://pvcreq.dsh.cms.gov/extension/1.0" schemaLocation="../.././XMLschemas/extension/ExtensionModel.xsd"/>
+  <xsd:element abstract="false" name="BatchHandlingServiceRequest" nillable="false" type="ext:BatchRequestType">
+    <xsd:annotation>
+      <xsd:documentation source="">A single transmission within a larger batch.</xsd:documentation>
+    </xsd:annotation>
+  </xsd:element>
+</xsd:schema>
