+++ lib/aca_entities/fdsh/pvc/h43/xsd/request/XMLschemas/subset/niem/wantlist.wlxml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="ASCII"?>
+<_2:WantList xmlns:_2="http://niem.gov/niem/wantlist/2" xmlns:nc="http://niem.gov/niem/niem-core/2.0" _2:product="NIEM" _2:release="2.1">
+  <_2:Element _2:isReference="false" _2:name="nc:DocumentBinary"/>
+  <_2:Element _2:isReference="false" _2:name="nc:DocumentFileName"/>
+  <_2:Element _2:isReference="false" _2:name="nc:DocumentSequenceID"/>
+  <_2:Element _2:isReference="false" _2:name="nc:DocumentBinary"/>
+  <_2:Element _2:isReference="false" _2:name="nc:DocumentFileName"/>
+  <_2:Element _2:isReference="false" _2:name="nc:DocumentSequenceID"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:AmountType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:BinaryType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:DocumentType">
+    <_2:ElementInType _2:isReference="false" _2:maxOccurs="1" _2:minOccurs="1" _2:name="nc:DocumentFileName"/>
+    <_2:ElementInType _2:isReference="false" _2:maxOccurs="1" _2:minOccurs="1" _2:name="nc:DocumentSequenceID"/>
+  </_2:Type>
+  <_2:Type _2:isRequested="false" _2:name="nc:NonNegativeDecimalSimpleType">
+    <_2:Facet _2:facet="minExclusive" _2:value="1"/>
+  </_2:Type>
+  <_2:Type _2:isRequested="false" _2:name="nc:NonNegativeDecimalType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:NumericType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:PercentageSimpleType">
+    <_2:Facet _2:facet="maxInclusive" _2:value="1"/>
+    <_2:Facet _2:facet="minExclusive" _2:value="0"/>
+  </_2:Type>
+  <_2:Type _2:isRequested="false" _2:name="nc:PercentageType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:PersonNameTextType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:ProperNameTextType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:QuantityType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:TextType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:AmountType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:BinaryType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:DocumentType">
+    <_2:ElementInType _2:isReference="false" _2:maxOccurs="1" _2:minOccurs="1" _2:name="nc:DocumentFileName"/>
+    <_2:ElementInType _2:isReference="false" _2:maxOccurs="1" _2:minOccurs="1" _2:name="nc:DocumentSequenceID"/>
+  </_2:Type>
+  <_2:Type _2:isRequested="false" _2:name="nc:NonNegativeDecimalSimpleType">
+    <_2:Facet _2:facet="minExclusive" _2:value="1"/>
+  </_2:Type>
+  <_2:Type _2:isRequested="false" _2:name="nc:NonNegativeDecimalType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:NumericType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:PercentageSimpleType">
+    <_2:Facet _2:facet="maxInclusive" _2:value="1"/>
+    <_2:Facet _2:facet="minExclusive" _2:value="0"/>
+  </_2:Type>
+  <_2:Type _2:isRequested="false" _2:name="nc:PercentageType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:PersonNameTextType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:ProperNameTextType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:QuantityType"/>
+  <_2:Type _2:isRequested="false" _2:name="nc:TextType"/>
+</_2:WantList>
